Avoid calling SaveHelper with size_t (MK)
git-svn-id: https://svn.code.sf.net/p/ourorgan/svn/trunk@600 69509848-2a73-49d5-a893-c4d8c934025a

+++ src/grandorgue/GOrgueDivisional.cpp
@@ -103,9 +103,9 @@ void GOrgueDivisional::Save(IniFileConfig& cfg, bool prefix)
 
 	GOrguePushbutton::Save(cfg, prefix);
 
-	cfg.SaveHelper(prefix, m_group, wxT("NumberOfStops"), m_Stops.size());
-	cfg.SaveHelper(prefix, m_group, wxT("NumberOfCouplers"), m_Couplers.size());
-	cfg.SaveHelper(prefix, m_group, wxT("NumberOfTremulants"), m_Tremulants.size());
+	cfg.SaveHelper(prefix, m_group, wxT("NumberOfStops"), (int)m_Stops.size());
+	cfg.SaveHelper(prefix, m_group, wxT("NumberOfCouplers"), (int)m_Couplers.size());
+	cfg.SaveHelper(prefix, m_group, wxT("NumberOfTremulants"), (int)m_Tremulants.size());
 
 	for (i = 0; i < m_Stops.size(); i++)
 	{
